Refactor completed.
