+++ client/src/contexts/index.jsx
@@ -1,1 +1,2 @@
 export { AuthProvider, useAuth } from "./Authentication";
+export { GardenProvider, useGarden } from "./Garden";
